adding jshint to make, fixing all jshint warnings, separating Email class from SESClient for easier maintenance and alphabetizing all functions - all tests pass
