+++ package.json
@@ -1,9 +1,9 @@
 {
   "author": "Aaron Heckmann <aaron.heckmann+github@gmail.com>",
-  "contributors" : [
+  "contributors": [
     "Mark Stosberg <mark@rideamigos.com> http://mark.stosberg.com"
   ],
-  "license" : "MIT",
+  "license": "MIT",
   "description": "Amazon SES (sendmail) for node",
   "keywords": [
     "amazon",
@@ -26,6 +26,7 @@
     "xml2js": "0.4.*"
   },
   "devDependencies": {
+    "jshint": "^2.8.0",
     "mocha": "*"
   },
   "optionalDependencies": {},
