+++ test/index.js
@@ -1,11 +1,12 @@
+'use strict';
 
 // use mocha for test
 
 var assert = require('assert')
   , ses = require('../')
-  , crypto = require('crypto')
+  , crypto = require('crypto');
 
-function create () {
+function create (){
   return new ses.Email({
       key: 'key'
     , secret: 'secret'
@@ -22,43 +23,52 @@ function create () {
 }
 
 describe('node-ses', function(){
-  it('should have createClient and Email exports', function () {
+  it('should have createClient and Email exports', function (){
     assert.equal('function', typeof ses.createClient);
     assert.equal('function', typeof ses.Email);
   });
+
   it('should have a version', function(){
     assert(ses.version);
-  })
-})
+  });
+});
 
 describe('createClient', function(){
   var client = ses.createClient({
       key: 'mykey'
     , secret: 'mysecret'
   });
+
   it('should have a key', function(){
     assert.ok(client.key);
   });
+
   it('should have a secret', function(){
     assert.ok(client.secret);
   });
+
   it('should have an algorithm', function(){
     assert.ok(client.algorithm);
   });
+
   it('algorithm should default to SHA1', function(){
     assert.equal(client.algorithm, 'SHA1');
   });
+
   it('algorithm should be overiddable', function(){
     var client = ses.createClient({ algorithm: 'different', key: 1, secret: 2 });
     assert.equal(client.algorithm, 'different');
   });
+
   it('should have an amazon property', function(){
     assert.ok(client.amazon);
   });
+
   it('amazon should default correctly', function(){
     var amazon = 'https://email.us-east-1.amazonaws.com';
     assert.equal(client.amazon, amazon);
   });
+
   it('amazon should be overiddable', function(){
     var client = ses.createClient({
         amazon: 'http://www.google.com'
@@ -67,26 +77,33 @@ describe('createClient', function(){
     });
     assert.equal(client.amazon, 'http://www.google.com');
   });
+
   it('should require a key', function(){
     try {
-      ses.createClient()
+      ses.createClient();
     } catch (err) {
-      if (!/key is required/.test(err)) throw err;
+      if (!/key is required/.test(err)) {
+        throw err;
+      }
     }
   });
+
   it('should require a secret', function(){
     try {
-      ses.createClient({ key: 1 })
+      ses.createClient({ key: 1 });
     } catch (err) {
-      if (!/secret is required/.test(err)) throw err;
+      if (!/secret is required/.test(err)) {
+        throw err;
+      }
     }
   });
+
   describe('sendemail', function(){
     it('should be a function', function(){
-      assert.equal('function', typeof client.sendemail)
-    })
-  })
-})
+      assert.equal('function', typeof client.sendemail);
+    });
+  });
+});
 
 describe('Email', function(){
   var email = new ses.Email({
@@ -105,52 +122,60 @@ describe('Email', function(){
 
   it('should have key', function(){
     assert.equal('key', email.key);
-  })
+  });
+
   it('should have secret', function(){
     assert.equal('secret', email.secret);
-  })
+  });
+
   it('should have algorithm', function(){
     assert.equal('algo', email.algorithm);
-  })
+  });
+
   it('should have from', function(){
     assert.equal('from', email.from);
-  })
+  });
+
   it('should have subject', function(){
     assert.equal('subject', email.subject);
-  })
+  });
+
   it('should have message', function(){
     assert.equal('message', email.message);
-  })
+  });
+
   it('should have altText', function(){
     assert.equal('alt', email.altText);
-  })
+  });
 
   describe('#extractRecipient', function(){
     it('should default to an array', function(){
       email.extractRecipient({ to: 'hi' }, 'from');
       assert(Array.isArray(email.from));
       assert.equal(0, email.from.length);
-    })
+    });
+
     it('should convert to an array', function (){
       email.extractRecipient({ to: 'hi' }, 'to');
       assert(Array.isArray(email.from));
       assert.equal(1, email.to.length);
       assert.equal('hi', email.to[0]);
-    })
+    });
+
     it('should leave arrays untouched', function(){
       email.extractRecipient({ yep: ['hi'] }, 'yep');
       assert(Array.isArray(email.yep));
       assert.equal(1, email.yep.length);
       assert.equal('hi', email.yep[0]);
-    })
-  })
+    });
+  });
 
   describe('#addMessage', function(){
     var email = create();
 
     it('should be a function', function(){
       assert.equal('function', typeof email.addMessage);
-    })
+    });
 
     it('should add subject,message,altText', function(){
       var msg = email.addMessage({});
@@ -167,14 +192,14 @@ describe('Email', function(){
       assert.equal(msg['Message.Body.Html.Charset'], 'UTF-8');
       assert.equal(msg['Message.Body.Text.Data'], 'alt');
       assert.equal(msg['Message.Body.Text.Charset'], 'UTF-8');
-    })
-  })
+    });
+  });
 
   describe('#addDestination', function(){
     var email = create();
     it('should be a function', function(){
       assert.equal('function', typeof email.addDestination);
-    })
+    });
 
     it('should add Destination data', function(){
       var msg = email.addDestination({});
@@ -189,54 +214,54 @@ describe('Email', function(){
       assert.equal(msg['Destination.CcAddresses.member.1'], 'cc@example.com');
       assert.equal(msg['Destination.BccAddresses.member.1'], 'bcc@example.com');
       assert.equal(msg['Destination.BccAddresses.member.2'], 'yoyo@example.com');
-    })
-
-
-
-  })
+    });
+  });
 
   describe('#addReplyTo', function(){
     var email = create();
+
     it('should be a function', function(){
       assert.equal('function', typeof email.addReplyTo);
-    })
+    });
+
     it('should add replyTo data', function(){
       var msg = email.addReplyTo({});
       assert(msg['ReplyToAddresses.member.1']);
       assert.equal(msg['ReplyToAddresses.member.1'], 'bcc@example.com');
-    })
-  })
+    });
+  });
 
   describe('#signature', function(){
     var email = create();
-    email.algorithm = 'SHA1'
+    email.algorithm = 'SHA1';
 
-    var sig = crypto.createHmac(email.algorithm.toLowerCase(), email.secret)
-                .update(email.date)
-                .digest('base64');
+    var sig = crypto
+      .createHmac(email.algorithm.toLowerCase(), email.secret)
+      .update(email.date)
+      .digest('base64');
 
     it('should be a function', function(){
       assert.equal('function', typeof email.signature);
-    })
+    });
 
     it('should compute the base64 hmac', function(){
       assert.equal(sig, email.signature());
-    })
+    });
 
     it('should return the same value if called > 1', function(){
       assert.equal(sig, email.signature());
       assert.equal(sig, email.signature());
       assert.equal(email._signature, email.signature());
-    })
-  })
+    });
+  });
 
   describe('#headers', function(){
     var email = create();
-    email.algorithm = 'SHA1'
+    email.algorithm = 'SHA1';
 
     it('should be a function', function(){
       assert.equal('function', typeof email.headers);
-    })
+    });
 
     it('should add headers', function(){
       var h = email.headers({});
@@ -245,22 +270,25 @@ describe('Email', function(){
       assert(/AWSAccessKeyId=/.test(h['X-Amzn-Authorization']));
       assert(/Algorithm=/.test(h['X-Amzn-Authorization']));
       assert(/Signature=/.test(h['X-Amzn-Authorization']));
-    })
-  })
+    });
+  });
 
   describe('#validate', function(){
     var email = create();
+
     it('should be a function', function(){
       assert.equal('function', typeof email.validate);
-    })
+    });
+
     it('should pass', function(){
       assert.equal(undefined, email.validate());
-    })
+    });
+
     it('should fail with To, Cc or Bcc is required', function(){
       email.to = [];
       email.cc = [];
       email.bcc = [];
-      assert.equal("To, Cc or Bcc is required", email.validate());
+      assert.equal('To, Cc or Bcc is required', email.validate());
 
       email.cc = ['works'];
       assert.equal(undefined, email.validate());
@@ -272,38 +300,42 @@ describe('Email', function(){
 
       email.to = ['works'];
       assert.equal(undefined, email.validate());
-    })
+    });
+
     it('should fail with From is required', function(){
       delete email.from;
-      assert.equal("From is required", email.validate());
+      assert.equal('From is required', email.validate());
       email.from = null;
-      assert.equal("From is required", email.validate());
+      assert.equal('From is required', email.validate());
       email.from = '';
-      assert.equal("From is required", email.validate());
+      assert.equal('From is required', email.validate());
       email.from = 'yup@asdf.com';
       assert.equal(undefined, email.validate());
-    })
+    });
+
     it('should fail with Subject is required', function(){
       delete email.subject;
-      assert.equal("Subject is required", email.validate());
+      assert.equal('Subject is required', email.validate());
       email.subject = null;
-      assert.equal("Subject is required", email.validate());
+      assert.equal('Subject is required', email.validate());
       email.subject = '';
-      assert.equal("Subject is required", email.validate());
+      assert.equal('Subject is required', email.validate());
       email.subject = 'spammer';
       assert.equal(undefined, email.validate());
-    })
-  })
+    });
+  });
 
   describe('#data', function(){
     var email = create();
+
     it('should be a function', function(){
       assert.equal('function', typeof email.data);
-    })
+    });
+
     it('should contain the formatted emails data', function(){
       var d = email.data({});
-      assert(d.Action)
-      assert.equal(d.Action, 'SendEmail')
+      assert(d.Action);
+      assert.equal(d.Action, 'SendEmail');
       assert(d.AWSAccessKeyId);
       assert.equal(email.key, d.AWSAccessKeyId);
       assert(d.Signature);
@@ -328,29 +360,32 @@ describe('Email', function(){
       assert(d['Message.Body.Text.Data']);
       assert(d['Message.Body.Text.Charset']);
       assert(d['ReplyToAddresses.member.1']);
-    })
-  })
+    });
+  });
 
   describe('#send', function(){
     var email = create();
-    email.algorithm = 'SHA1'
+
+    email.algorithm = 'SHA1';
     email.amazon = ses.amazon;
+
     it('should be a function', function(){
       assert.equal('function', typeof email.send);
-    })
+    });
+
     it('should callback an error', function(done){
       this.timeout(5000);
       var calledTimes = 0;
       email.send(function (err) {
         calledTimes++;
-        assert.equal(calledTimes,1,"callback was called only once");
+        assert.equal(calledTimes,1,'callback was called only once');
         assert(err);
         assert(err.Message);
         assert(/^The security token included in the request is invalid/.test(err.Message));
         // Wait to see if the code is accidentally going to run the test before declaring done.
         setTimeout(done,1000);
       });
-    })
+    });
 
     if (process.env.NODE_SES_KEY && process.env.NODE_SES_SECRET) {
       it('should succeed', function(done){
@@ -358,6 +393,7 @@ describe('Email', function(){
             key: process.env.NODE_SES_KEY
           , secret: process.env.NODE_SES_SECRET
         });
+
         client.sendemail({
             from: 'noreply@learnboost.com'
           , subject: 'testing node-ses'
@@ -372,9 +408,9 @@ describe('Email', function(){
           assert(data);
           done();
         });
-      })
+      });
     } else {
       console.error('You are not testing with Amazons SES service');
     }
-  })
-})
+  });
+});
